\name{xydat}
\alias{xydat}
\docType{data}
\title{
xydat
}
\description{
A dataset containing 61 xy-points 
}
\usage{data("xydat")}
\format{
  A data frame with 61 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\examples{
data(xydat)
}
\keyword{datasets}
